requirements/base.txt
# Configuration
django-environ==0.4.0

# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# images
Pillow==3.2.0
sorl-thumbnail==12.3

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2

# For the persistance stores
mysqlclient==1.3.7

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Useful things
# Time zones support
pytz==2016.3

# Your custom requirements go here
django-guardian==1.4.4
django-mailbox==4.4.8
claw==1.3.0
django-filter==0.13.0
django-autocomplete-light==2.3.3 # rq.filter: <3.0.0
https://github.com/ad-m/django-flatpages-tinymce/archive/patch-1.zip
https://github.com/ad-m/django-atom/archive/v0.11.0.zip
django-tinymce==2.3.0
django-tinycontent==0.5.1
icalendar==3.9.2
html2text==2016.4.2
pyGravatar==0.0.6
djmail==0.12.0.0.post1
django-bootstrap-pagination==1.6.1
raven==5.13.0
https://bitbucket.org/ad_m/django-bleach/get/6011e84.tar.gz
cached_property==1.3.0
django-extra-views==0.7.1
